requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-pptx==0.6.23
google-api-python-client==2.108.0
google-auth==2.25.2
requests==2.31.0
python-dotenv==1.0.0
playwright==1.48.0
beautifulsoup4==4.12.3
